requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-filter
